requirements/base.txt
django==4.1.1
python-dotenv==1.0.0
psycopg2-binary==2.9.6
requests==2.31.0
mongoengine==0.27.0
django-cors-headers==4.3.1
djangorestframework==3.14.0